requirements.txt
# Core ML and NLP libraries
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
jsonschema>=4.19.0

# Evaluation and metrics
nltk>=3.8.0
rouge-score>=0.1.2
bert-score>=0.3.13
spacy>=3.7.0

# Utilities
tqdm>=4.65.0
wandb>=0.15.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Web interface (optional)
streamlit>=1.28.0
gradio>=3.50.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0